isAcceptedTerms

+++ app/main.dev.js
@@ -16,6 +16,8 @@ import log from 'electron-log';
 import { exec } from 'child_process';
 import path from 'path';
 import fs from 'fs';
+import bs58 from 'bs58';
+import crypto from 'crypto';
 import MenuBuilder from './menu';
 
 export default class AppUpdater {
@@ -57,9 +59,13 @@ function runNodeProcess(pass: string): Promise<void> {
       const nodePath = path.resolve(__dirname, '../node/');
       const passFile = `${nodePath}/pass`; // todo config
       fs.writeFileSync(passFile, pass || '');
+      const encoded = bs58.encode(crypto.randomBytes(16));
+      const tokenFile = `${nodePath}/api_token.txt`; // todo config
+      fs.writeFileSync(tokenFile, encoded);
       nodeProcess = exec(`${nodePath}/stegos`, { cwd: nodePath }, err => {
         if (err) {
-          fs.unlinkSync(passFile);
+          if (fs.existsSync(passFile)) fs.unlinkSync(passFile);
+          if (fs.existsSync(tokenFile)) fs.unlinkSync(tokenFile);
           reject(err);
         }
       });
@@ -69,9 +75,13 @@ function runNodeProcess(pass: string): Promise<void> {
           str.includes('Network endpoints:') ||
           str.includes('Adding node from seed pool')
         ) {
+          if (fs.existsSync(passFile)) fs.unlinkSync(passFile);
+          if (fs.existsSync(tokenFile)) fs.unlinkSync(tokenFile);
           resolve();
         }
         if (str.includes('Invalid password:')) {
+          if (fs.existsSync(passFile)) fs.unlinkSync(passFile);
+          if (fs.existsSync(tokenFile)) fs.unlinkSync(tokenFile);
           reject(new Error('Invalid password'));
         }
       });
